Fixes issue #1022 - Add module-info.java to all our JARs (#1040)
* JPMS - Initial module-info #1022

The minimal module-infos of the modules.
It can startup Piranha Server.
It doesn't build the tests yet (-Dmaven.test.skip)
Modules skipped: MicroProfile OmnifacesJWT, Piranha Micro and
Arquillian Server

* JPMS - Limit scope of ServiceLoader and reorder requires #1022

Don't provide services that currently aren't exposed, i.e. don't
have a service file

* JPMS - Add module info to some wars #1022

The compiler plugin requires that all modules have a module-info.java
or a Automatic-Module-Name. It's missing some modules though.

* JPMS - Fix Http module tests compilation #1022

Producing both a jar and a test-api doesn't go well with the JPMS.
Extract the base class of http tests to a new module (test-api).

* Revert "JPMS - Add module info to some wars #1022"

This reverts commit ef87cd4a9d4634e6a6a09a687abca7aca5e53ccf.

* JPMS - Disable aggregate javadoc for now #1022

Enable it again after have completed the tests are passing

* JPMS - Bump test-api version #1022

* JPMS - Makes the compilation with tests go further #1022

* JPMS - Jasper needs to Search for classes from Modulepath #1022

As we are going modular, we need to tell the Jasper search the classes
from modulepath, not classpath

* JPMS - Fix building with tests and Javadoc #1022

* Revert "JPMS - Disable aggregate javadoc for now #1022"

This reverts commit 7149fad2b275725ca2fb743ffde90998f26a7c2d.

* JPMS - Fixes Javadoc aggregator-jar build #1022

Change the package of micro-core module to avoid split packages
with the micro module and add module-info to both micro and
arquillian-server modules.

* JPMS - Include module-info in the shaded jar #1022

* JPMS - Add missing license header #1022