Add the debug option and verbose (foreground) (#1949)
* Add the debug option and verbose (foreground)

I am proposing this PR, so we can debug and do foreground run (or start?) easily. I took "suspend" and "verbose" names from some other projects, so I think the user is familiar with it.

* Update start.sh

Adding --run option to start in background with a wait

* Update run.sh

Make sure all options in start.sh also apply when using run.sh

* Update start.sh

Touch the pid before the command, since there is some loop looking for it in the cmd, and the pid is otherwise created after it. We get a race here.

* Update server/src/main/zip/bin/run.sh

Co-authored-by: Thiago Henrique Hüpner <thihup@gmail.com>
Co-authored-by: Arjan Tijms <arjan.tijms@gmail.com>